+++ CHANGES.md
@@ -4,7 +4,7 @@
 - Support IntBuffer in GL20
 - Update teavm to 0.10.0-dev-1
 - Update libgdx to 1.12.1
-
+- Remove obfuscate from build config. Use TeaVMTool.
 
 [1.0.0-b7]
 - Remove jParser